[SEMI-MODULAR] Mining toys part 2: PKA edition (#1332)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Continuing the port of Monkestation's mining tools
(https://github.com/Monkestation/Monkestation2.0/pull/489)
This adds SIX new Proto-kinetic accelerator variants, available for 1250
mining points:
The Pistol, dealing a measly 10 damage per shot but sporting 2x the mod
capacity.
The repeater, sporting 75 mod capacity and 20 damage per shot, but able
to fire thrice before needing recharging.
The shotgun, sporting 75 mod capacity and firing three 20-damage
pellets.
The shockwave, trading range for a 3x3 AoE around the wielder. 75 mod
capacity.
The grenade launcher, firing the same bombs as the mining modsuit. Does
not accept modkits.
The railgun, firing an impressive 100 damage per shot, which pierces
through mobs. Recharges slower than other PKAs, and does not accept
modkits.

The railgun and grenade launcher are equipped with a special Wastes
firing pin, preventing their use outside of Lavaland/icemoon.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Variety is the spice of life. This adds a lot of alternatives to the
classic PKA. Paired with the crusher variants, this brings more options
in terms of playstyle and flair for mining players.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/Bubberstation/Bubberstation/assets/21011148/68bd2196-848b-40f6-abb9-7c6f18c9b200)
Lookit them
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:VrazzleDazzle
add: Six new PKA variants, now in a mining vendor near you.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>