Disables Station Blackout and Random Machine Languages Station Traits (#1399)
## About The Pull Request

Disables the Station Blackout and Random Machine Language station
traits.

## Why It's Good For The Game

Station Blackout would kill all the lights on station. This was awful,
especially if there was no janitor.

Random Machine Language would give a random language to every machine
that speaks and every bot. This was annoying since it affected cryo
announcements and arrivals announcements.

## Proof Of Testing

If it compiles, it werks.

## Changelog


:cl: BurgerBB
del: Disables Station Blackout and Random Machine Languages Station
Traits
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->