[MODULAR] Dauntless map tweaks and bounty pad (#1374)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR tweaks the following things for dauntless lava ruins:
* Removes improperly copied decals under hull tiles
* Adjusts the recreational area for the brig
* Adds many exterior cameras to the dauntless, and removes many interior
ones (except for those in brig & interrogation)
* Adds a maxcap of 1MW to the powerator
* Adds a modified pirate bounty pad that allows the Dauntless crew to
sell off goods for credits
* Adds a produce order console to the kitchen which is linked to the
Interdyne budget account
* Adds a new area variable called `ignore_weather_sfx` which bypasses
weather sfx (this is applied to the Dauntless)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

This PR removes the possibility for the AI to go snooping in the
Dauntless super sekret base, and can only view the exterior. It also
tweaks areas such as dorms and brig to be easier on both the eyes and
for movement.

The new bounty pad will allow players to use alterative forms of money,
be it selling off the ore from netrunner gaming den, or other making
goods for credits. This means they are not relying on using the
powerator (which has been capped at 1MW).

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

I have ran this map on my local server and confirm it works.

The bounty pad works and I have ensured that there are enough sanity
checks for it to not crash or break anything. It should not affect on
station cargo prices.


https://github.com/Bubberstation/Bubberstation/assets/80724828/f8b5210a-4664-4c4b-8c6e-be17385f41c2

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Dauntless - Add exterior cameras to dauntless, removed many
interior ones
fix: Dauntless - Removed incorrectly placed decals under hull objects
and tweaked brig/dorms
balance: Interdyne powerator is now maxcapped to 1MW
add: Dauntless - Added new bounty pad and bounty console
add: Dauntless - Added new produce order console
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>