[NOT MODULAR][TM] Nerfs Bluespace crystals (#1402)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a time to use bluespace crystals of a few seconds

Honestly this might be doable entirely without the uncalibrated delimb
chance. We'll see how it tests out.

This one actually isn't a salt PR I'm just going off of how annoying it
is as a ghostmain to watch people teleport away scott free when the odds
aren't in their favor. Also this
[suggestion's](https://discord.com/channels/1059199070016655462/1071095123145924679/1230587434773778462)
results imply that players are kind of tired of it and I generally
believe that if an overwhelming majority of respondents want something
maybe that something isn't a bad idea.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Now that arcmining is creating infinite bluespace crystals we are seeing
antags get literal stacks of them and be uncatchable the entire shift.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

(Just pretend there's no delimbing and screaming because that's been
removed)

![image](https://github.com/Bubberstation/Bubberstation/assets/26744576/4c8ee995-5972-471e-a981-c4e9829688dc)

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Swiftfeather
balance: Makes bluespace crystal teleportation take time.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>