Automapped NTC office updates (#1366)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR updates our automapped NTC templates with some minor fixes, but
most importantly, adds unrestricted access helpers from the inside of
the office, meaning people inside of the office can freely exit.

I do have a question I'd like to ask for this PR, should the bedroom
also include an unrestricted helper? It currently does but if you want
to prevent, say, bluespacing into the office and stealing stuff, we
could just take the inner bedroom helper off.

Also this means the Blueshield won't be able to escape my ERP scene, so
double positive
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->
## Why It's Good For The Game
This means that if people enter the NTC's office, they can exit on their
own. The office is unique in that it requires an access no one other
than the NTC can acquire, so this is a bit of a bandaid for that.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
Tested all the different access helpers, they work as intended :)
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: NTC office bugs have received some minor fixes
add: It is now free to exit the NTC office, even if the outer access is
not free
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

Co-authored-by: Swift <jackwars4@gmail.com>