[Non-Modular Because Globalvars suck ass] Mirror-Nova: "Girl Dinner: Adds Mariinsky Ballet Company Chassis" (#1169)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
### This will be a 1:1-ish paste of the Nova PR's description. I noticed
that the actual pull from Nova did not edit the RD's labcoat like it
says it does.

This PR adds:

A new synth chassis, Mariinsky Ballet Company. Sprite inspired by
Signalis and Atomic Heart.

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

More synth character customization. Peg legs have been requested for a
long time, especially in the wake of Signalis' release. The sprite is
compatible with most hairstyles and clothing.

## Proof Of Testing

### Nova's Screenshots

![304930570-1a99a916-0be7-4f2e-87d2-98656448daa7](https://github.com/Bubberstation/Bubberstation/assets/5521322/5c3716aa-d475-4dcf-99c8-f7db7a02316f)

![304930656-3c8a2637-8129-400b-b9f7-d262355cce6c](https://github.com/Bubberstation/Bubberstation/assets/5521322/5bf6788d-c465-47a3-906f-32f58e217c42)

### My screenshot to show it actually loaded finally

![2024-02-18_00-23-40](https://github.com/Bubberstation/Bubberstation/assets/5521322/cb4b4f4f-45ad-4f28-b0dc-2f404da0cc8f)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Parts from a discontinued synth chassis line from the Mariinsky
Ballet Company is making rounds on the black market - now available to
install for synths and chromed up organics!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>