Adds the Wizard Die of Fate, plus the random event that spawns it. Roll a 20, and become a wizard. Roll a 1, and fucking die. (#1414)
## About The Pull Request

Adds the Wizard Die of Fate. Roll a 20, and become a wizard. Roll a 1,
and fucking die (not dust).
If someone becomes a wizard, the die is deleted, if not, it will
teleport to a random area on the station. The dice has 20 uses, and will
delete itself if it has no more uses left.

The dice is obtained by the event that goes with it. The event will
inform crew that a magical dice has spawned somewhere on station and you
should touch it. There is no indication of where it spawns.

## Why It's Good For The Game

This is a way to add wizard naturally in the game without adding another
disgusting ghost role for antag fishers to idle for. Wizard is currently
pretty balanced. If not I can easily just salt PR away anyone who abuses
spells for anything I missed.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/5cbc74bf-f3fc-4351-93da-bd054b2dbc61)

## Changelog

:cl: BurgerBB
add: Adds the Wizard Die of Fate, plus the random event that spawns it.
Roll a 20, and become a wizard. Roll a 1, and fucking die.
/:cl: