some hairs and a cool mot wing (#1440)
## About The Pull Request

adds 3 hairstyles, 2 of which are edits of existing hairstyles (bedhead
and spicy) and a third one that is a cool shortish horse mane

also adds a cool wing (citheronia) from sojourn station, now in a
colorable version and with the front trimmed a little so your leggies
and lower torso can show when facing south!


## Why It's Good For The Game

💅💅💅💅💅

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/129433703/190899f1-5ebc-437b-8fb2-4304fe564c46)

![image](https://github.com/Bubberstation/Bubberstation/assets/129433703/042f42c2-9002-4391-bd18-7aa6df0e1a9b)

![image](https://github.com/Bubberstation/Bubberstation/assets/129433703/babcbd8a-cfde-49cc-a73f-bbcc9a26851c)

![image](https://github.com/Bubberstation/Bubberstation/assets/129433703/2ba06ab3-6b30-4e50-9387-3f9f2b835e3f)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: 3 new hairstyles! Bedhead (+hairline) that fixes the hairline of
Bedhead, an alt color version of Spicy and a brand new, short horse
mane.
add: A colorable version of the Citheronia wings from Sojourn Station,
also with the front part trimmed so the legs and torso can show a
little.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>