Makes Ashwalker/Lizard names use a special name generator (#1518)
## About The Pull Request

Makes Ashwalker/Lizard names use a special generator, which either makes
its own name based on consonant/vowel pairs, or with the infamous
Verbs-The-Noun format.

Consonant/vowel generation will take up 80% of random names.
Verbs-The-Noun will take up 20% of random names.

Uses the ic/slur filter as well in case it generates Sus-Amongus or
something hilarious.

## Why It's Good For The Game

This is basically a test PR that I will eventually(tm) make for /tg/. I
just want to know what people think of it first.

The reason why I made this is because I made the mistake of using the
standard /tg/ generator for one of my static characters and it's
entirely possible for someone else to get the same name since there are
limited pickings (1 in 26569 for a female lizardperson).

## Proof Of Testing

Using the new debug verb provided:


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/ad8195ac-4ce8-4236-9c56-3e27d878c99d)


Example in the character creator:


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/85039df4-5321-47b7-aa1c-59cb37ad7585)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
add: Makes Ashwalker/Lizard names use a special name generator
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>