Moonstation Mega Multi-Z Update (#1541)
![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/466a33d5-9bd0-4afd-bc48-eafa3502118c)


## About The Pull Request

Destroying floor tiles on the Moonstation upper level can lead to a
chasm that drops you down to the middle cave level. Note that you can't
get dropped down to lavaland.
The middle cave level is now breathable air; the same as the surface.
Several areas were tweaked and some areas were added to include
additional ladders that lead to the surface.
The prison is now "harder" to escape, with an already-existing chasm
outside the windows.

## Why It's Good For The Game

It might actually not be because it means people might die to the mobs
under there, and then get upset. Lets find out.

Regardless it looks pretty cool.

## Proof Of Testing

See screenshots.

## Ladder Routes


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/033a35b2-ed21-4242-b7ff-50706b7c212d)

Routes marked with * require additional access.

## Changelog

:cl: BurgerBB
add: Destroying floor tiles on the Moonstation upper level can lead to a
chasm that drops you down to the middle cave level. Note that you can't
get dropped down to lavaland.
/:cl:

---------

Co-authored-by: Swift <jackwars4@gmail.com>