Reduces nightmare's burn weakness (#1455)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
[IDED PR]
Currently nightmare has a 1.5 burn mod, and TG has recently buffed laser
damage. This means that nightmares really suffer as a result.
This PR reduces the burn mod to 1.2, so that nightmares wont did in less
than a second to lasers.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Lasers deal enough damage to crit a nightmare in 2 volleys if you have
lasers in both hands, and the fire speed is fast enough that there's not
enough time to react. Usually it's not so much issue for most other
antags, as they can wear actual armor. However, nightmares cannot, and
lasers can hit them even in the dark, as they generate light sources.
That's not to mention the other high damage sources of burns that
security has. (Laser shells, custom lasers)
Main issue here is a nightmare has barely enough chance to react to the
damage, and is a fragile antag as they're very unlikely to revive if
dead, and their main benefit to surviving is the self-healing in the
dark, but cannot wear any armor.
This gives the nightmare a chance to react to burn damage, at the least
Note: Scattershot lasers take 2 shells to down a nightmare from point
blank, which is likely given they are a melee only and fight in 1-tile
maint corridors
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
Went ingame, shot at a nightmare, noticed they didn't fold
near-instantly
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Reduced nightmare's burn weakness to 1.2, from 1.5.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>