Nerfs borg tipping's stun from 60 seconds to just 20 seconds (#1529)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I love Silicons.

This PR makes it so instead of the agonizingly long 60 second timer on
righting yourself as a borg, you have a 20 second timer instead. It's
still a very long stun, but is not an entire minute in a game where you
can be killed in 3 seconds.
See below for how long an entire minute stun is.


https://github.com/Bubberstation/Bubberstation/assets/110273561/4fb15f99-4816-4106-b388-a787635052bc


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
There's long, and then there's too long. This is too long, and makes the
game worse for cyborg players without much tangible benefit.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
With & without using beep emote

https://github.com/Bubberstation/Bubberstation/assets/110273561/f20c3640-46c4-4d86-a618-384db338dbc8


https://github.com/Bubberstation/Bubberstation/assets/110273561/e1f718f7-a5fe-4f0a-8d97-fd6ce0af6c52

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: ReturnToZender
balance: The stun time on the tippable component for borgs has been
reduced to 20 seconds, from 60 seconds.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: Swift <jackwars4@gmail.com>